requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
datamodel-code-generator
sqlalchemy
asyncpg
pydantic[email]
psycopg2-binary
pytest
httpx
PyYAML
flask_bcrypt